Add terraform that manages deployment infrastructure. (#6)
* Add terraform setup and initial AWS configuration.

* Parameterize app and service ports.

* Use data source instead of variable for AMI ID.

* Attach cloud-config to the LaunchConfiguration.

* Use one subnet per AZ and fix some name errors.

* Use Ignition instead of deprecated cloud-config.

* Add .dockerignore.

Avoids adding the terraform code to the built container.

* Fix silent error on empty app_version.

Do not set the default for `app_version` so that Terraform explicitely
asks for it if undefined. This fixes a silent error when provisioning to
AWS, which was caused by having docker trying to pull from
`quay.io/mtpereira/testing-express-api:`, which is an invalid tag name.